orders table into its own component, memoized
